+++ react/components/index.js
@@ -1,3 +1,4 @@
 export * from "./atoms";
+export * from "./contexts";
 export * from "./molecules";
 export * from "./organisms";