+++ react/components/organisms/chat/chat.js
@@ -26,7 +26,10 @@ export class Chat extends PureComponent {
                     )
                 })
             ),
+            animateScrollBottom: PropTypes.bool,
             onNewMessage: PropTypes.func,
+            onScrollBottom: PropTypes.func,
+            onScroll: PropTypes.func,
             style: ViewPropTypes.style
         };
     }
@@ -36,7 +39,10 @@ export class Chat extends PureComponent {
             avatarUrl: undefined,
             username: undefined,
             messages: [],
+            animateScrollBottom: true,
             onNewMessage: () => {},
+            onScrollBottom: () => {},
+            onScroll: event => {},
             style: {}
         };
     }
@@ -59,10 +65,20 @@ export class Chat extends PureComponent {
     };
 
     scrollToEnd = () => {
-        this.scrollViewComponent.scrollToEnd({ animated: true });
+        this.scrollViewComponent.scrollToEnd({ animated: this.props.animateScrollBottom });
     };
 
     getInputValue = () => (this.input ? this.input.state.value || null : null);
+
+    onScroll = event => {
+        this.props.onScroll(event);
+        if (
+            event.nativeEvent.layoutMeasurement.height + event.nativeEvent.contentOffset.y >=
+            event.nativeEvent.contentSize.height
+        ) {
+            this.props.onScrollBottom();
+        }
+    };
 
     async _onNewMessage(message) {
         this.setState({ sendingMessage: true }, () => {
@@ -89,7 +105,7 @@ export class Chat extends PureComponent {
             username: this.props.username,
             message: this.getInputValue(),
             date: Date.now(),
-            attachments: [source]
+            attachments: source
         };
         await this._onNewMessage(message);
     };
@@ -123,6 +139,7 @@ export class Chat extends PureComponent {
                     style={styles.chatMessagesContainer}
                     ref={ref => (this.scrollViewComponent = ref)}
                     onContentSizeChange={this.scrollToEnd}
+                    onScroll={this.onScroll}
                 >
                     <View style={styles.chatMessagesContent}>
                         {this.props.messages.map((message, index) => {